Sources

+++ ADNews/AppDelegate.m
@@ -7,8 +7,6 @@
 //
 
 #import "AppDelegate.h"
-#import "ADCoreDataManager.h"
-
 #import "ADTabBarController.h"
 
 @interface AppDelegate ()
@@ -17,47 +15,33 @@
 
 @implementation AppDelegate
 
-
 - (BOOL)application:(UIApplication *)application didFinishLaunchingWithOptions:(NSDictionary *)launchOptions {
-    // Override point for customization after application launch.
     ADTabBarController *mainController = [[ADTabBarController alloc] init];
     self.window.rootViewController = mainController;
     [self.window makeKeyAndVisible];
     
-    
-    [[ADCoreDataManager shared] deleteAllObjectsForEntityName:NSStringFromClass([DBNews class])];
-    [[ADCoreDataManager shared] deleteAllObjectsForEntityName:NSStringFromClass([DBSource class])];
-    [[ADCoreDataManager shared] deleteAllObjectsForEntityName:NSStringFromClass([DBDocument class])];
+    [self checkResetStatus];
     
     return YES;
 }
 
-
-- (void)applicationWillResignActive:(UIApplication *)application {
-    // Sent when the application is about to move from active to inactive state. This can occur for certain types of temporary interruptions (such as an incoming phone call or SMS message) or when the user quits the application and it begins the transition to the background state.
-    // Use this method to pause ongoing tasks, disable timers, and invalidate graphics rendering callbacks. Games should use this method to pause the game.
-}
-
-
-- (void)applicationDidEnterBackground:(UIApplication *)application {
-    // Use this method to release shared resources, save user data, invalidate timers, and store enough application state information to restore your application to its current state in case it is terminated later.
-    // If your application supports background execution, this method is called instead of applicationWillTerminate: when the user quits.
-}
-
-
-- (void)applicationWillEnterForeground:(UIApplication *)application {
-    // Called as part of the transition from the background to the active state; here you can undo many of the changes made on entering the background.
-}
-
-
 - (void)applicationDidBecomeActive:(UIApplication *)application {
-    // Restart any tasks that were paused (or not yet started) while the application was inactive. If the application was previously in the background, optionally refresh the user interface.
+    [self checkResetStatus];
+    [[ADCoreDataManager shared] downloadNewsForType:DBNewsTypeTopHeadlines withOptions:nil];
 }
 
-
-- (void)applicationWillTerminate:(UIApplication *)application {
-    // Called when the application is about to terminate. Save data if appropriate. See also applicationDidEnterBackground:.
+- (void)checkResetStatus {
+    NSString *resetKey = @"ADNewsResetDB";
+    NSUserDefaults *userDefaults = [NSUserDefaults standardUserDefaults];
+    NSNumber *number = [userDefaults valueForKey:resetKey];
+    BOOL reset = [number boolValue];
+    if (reset) {
+        [[ADCoreDataManager shared] deleteAllObjectsForEntityName:NSStringFromClass([DBNews class])];
+        [[ADCoreDataManager shared] deleteAllObjectsForEntityName:NSStringFromClass([DBSource class])];
+        [[ADCoreDataManager shared] deleteAllObjectsForEntityName:NSStringFromClass([DBDocument class])];
+        
+        [userDefaults setBool:!reset forKey:resetKey];
+    }
 }
-
 
 @end
